fix: split bets.test.ts in two. Fix betpool problems
