+++ dapp/test/betPool.test.ts
@@ -1,18 +1,14 @@
 import { describe, test, expect, beforeEach, vi } from 'vitest';
 import BetPool from '../src/BetPool';
-import Game from '../src/Game';
 import { WalletApp } from "@deroll/wallet";
-import { ValidatorFunctionRunner } from "../src/validator";
 import { Bet, PlayerBet } from "../src/types";
-import { Hex } from 'viem';
-//@TODO need to split this file into 2 files
-// betPool.test.ts and Game.test.ts
+import { Hex, toHex } from 'viem';
 
 describe('BetPool', () => {
     let betPool: BetPool;
     let betPoolSingle: BetPool;
     let mockWallet: WalletApp;
-    let mockToken: Hex = '0x1234';
+    let mockToken: Hex = '0xf795b3D15D47ac1c61BEf4Cc6469EBb2454C6a9b';
 
     beforeEach(() => {
         // Mock WalletApp for testing without actual wallet operations
@@ -20,7 +16,7 @@ describe('BetPool', () => {
             transferERC20: vi.fn(),
         } as unknown as WalletApp;
 
-        betPool = new BetPool(["football", "basketball"],mockToken, mockWallet);
+        betPool = new BetPool(["football", "basketball"], mockToken, mockWallet);
         betPoolSingle = new BetPool(["football"], mockToken, mockWallet);
     });
 
@@ -107,7 +103,7 @@ describe('BetPool', () => {
     test('should transfer remaining funds to DAO when closing', () => {
         // Simulate some remaining funds
         const ERC20_TOKEN = "0xf795b3D15D47ac1c61BEf4Cc6469EBb2454C6a9b"; //this is the sunodo token a sample erc20 token
-        const POOL_ADDRESS = "0x01"; //this is a sample address for bet pool
+        const POOL_ADDRESS = toHex(1); //this is a sample address for bet pool
         const DAO_ADDRESS = "0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266"; //this is a sample address for the DAO
 
         betPool.fundsLocked = BigInt(1000);
@@ -127,74 +123,5 @@ describe('BetPool', () => {
         expect(mockWallet.transferERC20).not.toHaveBeenCalled();
     });
 
-
-});
-
-describe('Game', () => {
-    let game: Game;
-    let mockWallet: WalletApp;
-    let mockValidatorFunctionRunner: ValidatorFunctionRunner;
-    let mockToken: Hex = '0x1234';
-
-
-    beforeEach(() => {
-        mockWallet = {
-            transferERC20: vi.fn(),
-        } as unknown as WalletApp;
-
-        mockValidatorFunctionRunner = {
-            run: vi.fn().mockResolvedValue("football"),
-        } as unknown as ValidatorFunctionRunner;
-
-        game = new Game(["football", "basketball"], Date.now(), Date.now() + 3600000,mockToken, mockValidatorFunctionRunner, mockWallet);
-        vi.spyOn(game.betPool, 'payout');
-    });
-
-    test('should make a bet and update player bets', () => {
-        const bet: Bet = {
-            pick: "football",
-            tokenAddress: mockToken,
-            player: "0xPlayer",
-            amount: BigInt(500),
-            effectiveAmount: BigInt(0)
-        };
-
-        game.makeBet(bet);
-
-        const playerBets = game.playersBets.get("0xPlayer");
-        expect(playerBets).toBeTruthy();
-        expect(playerBets!.get(bet.pick)).toContainEqual(bet);
-    });
-
-
-    test('should handle game settlement correctly', async () => {
-        game.makeBet({ pick: "football", tokenAddress: mockToken, player: "0xPlayer", amount: BigInt(1000), effectiveAmount: BigInt(0) });
-
-        await game.settle("game data", "0xSignature");
-
-        // Verify the settlement logic processes the winning pick
-        expect(mockValidatorFunctionRunner.run).toHaveBeenCalled();
-        expect(game.betPool.payout).toHaveBeenCalled();
-    });
-
-    test('should manage bets across different picks independently', () => {
-        game.makeBet({ pick: "football", tokenAddress: mockToken, player: "0xPlayer1", amount: BigInt(500), effectiveAmount: BigInt(0) });
-        game.makeBet({ pick: "basketball", tokenAddress: mockToken, player: "0xPlayer2", amount: BigInt(600), effectiveAmount: BigInt(0) });
-
-        const footballBets = game.playersBets.get("0xPlayer1")!.get("football");
-        const basketballBets = game.playersBets.get("0xPlayer2")!.get("basketball");
-
-        expect(footballBets).toContainEqual(expect.objectContaining({ amount: BigInt(500) }));
-        expect(basketballBets).toContainEqual(expect.objectContaining({ amount: BigInt(600) }));
-    });
-
-    test('should handle invalid winning picks correctly', async () => {
-        game.makeBet({ pick: "football", tokenAddress: mockToken, player: "0xPlayer", amount: BigInt(1000), effectiveAmount: BigInt(0) });
-
-        mockValidatorFunctionRunner.run.mockResolvedValue("invalid_pick"); // Simulate an invalid winning pick
-        await game.settle("game data", "0xSignature");
-
-        expect(game.betPool.payout).toHaveBeenCalledWith("invalid");
-    });
 
 });
